+++ packages/client/modules/meeting/components/TemplateScaleValueItem.tsx
@@ -10,6 +10,7 @@ import {TemplateScaleValueItem_scale} from '~/__generated__/TemplateScaleValueIt
 import Icon from '../../../components/Icon'
 import {PALETTE} from '../../../styles/paletteV2'
 import {ICON_SIZE} from '../../../styles/typographyV2'
+import isSpecialPokerLabel from '../../../utils/isSpecialPokerLabel'
 import {TemplateScaleValueItem_scaleValue} from '../../../__generated__/TemplateScaleValueItem_scaleValue.graphql'
 import EditableTemplateScaleValueColor from './EditableTemplateScaleValueColor'
 import EditableTemplateScaleValueLabel from './EditableTemplateScaleValueLabel'
@@ -18,7 +19,7 @@ interface Props {
   isDragging: boolean
   scale: TemplateScaleValueItem_scale
   scaleValue: TemplateScaleValueItem_scaleValue
-  dragProvided: DraggableProvided
+  dragProvided?: DraggableProvided
 }
 
 const ScaleValueItem = styled('div')<{isHover: boolean, isDragging: boolean}>(
@@ -74,13 +75,14 @@ const TemplateScaleValueItem = (props: Props) => {
     submitMutation()
     RemovePokerTemplateScaleValueMutation(atmosphere, {scaleId, label}, {}, onError, onCompleted)
   }
+  const isSpecial = isSpecialPokerLabel(label)
   return (
     <ScaleValueItem
-      ref={dragProvided.innerRef}
-      {...dragProvided.dragHandleProps}
-      {...dragProvided.draggableProps}
+      ref={dragProvided?.innerRef}
+      {...dragProvided?.dragHandleProps}
+      {...dragProvided?.draggableProps}
       isDragging={isDragging}
-      isHover={isHover}
+      isHover={!isSpecial && isHover}
       onMouseOver={onMouseOver}
       onMouseOut={onMouseOut}
     >
@@ -93,9 +95,12 @@ const TemplateScaleValueItem = (props: Props) => {
           scaleValue={scaleValue}
         />
       </ScaleAndDescription>
-      <RemoveScaleValueIcon isHover={isHover} onClick={removeScaleValue}>
-        cancel
-      </RemoveScaleValueIcon>
+      {
+        !isSpecial &&
+        <RemoveScaleValueIcon isHover={isHover} onClick={removeScaleValue}>
+          cancel
+        </RemoveScaleValueIcon>
+      }
     </ScaleValueItem >
   )
 }