+++ packages/server/graphql/mutations/newMeetingCheckIn.ts
@@ -6,6 +6,7 @@ import NewMeetingCheckInPayload from '../types/NewMeetingCheckInPayload'
 import toTeamMemberId from 'parabol-client/utils/relay/toTeamMemberId'
 import standardError from '../../utils/standardError'
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
+import sendMeetingJoinToSegment from './helpers/sendMeetingJoinToSegment'
 
 export default {
   type: NewMeetingCheckInPayload,
@@ -34,11 +35,7 @@ export default {
     const subOptions = {mutatorId, operationId}
     const viewerId = getUserId(authToken)
     // AUTH
-    const meeting = await r
-      .table('NewMeeting')
-      .get(meetingId)
-      .default(null)
-      .run()
+    const meeting = await r.table('NewMeeting').get(meetingId).default(null).run()
     if (!meeting) return standardError(new Error('Meeting not found'), {userId: viewerId})
     const {endedAt, teamId} = meeting
     if (endedAt) return standardError(new Error('Meeting already ended'), {userId: viewerId})
@@ -48,13 +45,10 @@ export default {
 
     // RESOLUTION
     const meetingMemberId = toTeamMemberId(meetingId, userId)
-    await r
-      .table('MeetingMember')
-      .get(meetingMemberId)
-      .update({isCheckedIn})
-      .run()
+    await r.table('MeetingMember').get(meetingMemberId).update({isCheckedIn}).run()
 
     const data = {meetingId, userId}
+    sendMeetingJoinToSegment(userId, meeting)
     publish(SubscriptionChannel.MEETING, meetingId, 'NewMeetingCheckInPayload', data, subOptions)
     return data
   }