+++ packages/server/graphql/mutations/pokerSetFinalScore.ts
@@ -13,11 +13,12 @@ import updateStage from '../../database/updateStage'
 import AtlassianServerManager from '../../utils/AtlassianServerManager'
 import {getUserId, isTeamMember} from '../../utils/authorization'
 import getJiraCloudIdAndKey from '../../../client/utils/getJiraCloudIdAndKey'
-import makeAppLink from '../../utils/makeAppLink'
+import makeAppURL from 'parabol-client/utils/makeAppURL'
 import makeScoreJiraComment from '../../utils/makeScoreJiraComment'
 import publish from '../../utils/publish'
 import {GQLContext} from '../graphql'
 import PokerSetFinalScorePayload from '../types/PokerSetFinalScorePayload'
+import appOrigin from '../../appOrigin'
 
 const pokerSetFinalScore = {
   type: GraphQLNonNull(PokerSetFinalScorePayload),
@@ -117,11 +118,8 @@ const pokerSetFinalScore = {
       )
       const fieldName = dimensionField?.fieldName ?? SprintPokerDefaults.JIRA_FIELD_COMMENT
       if (fieldName === SprintPokerDefaults.JIRA_FIELD_COMMENT) {
-        const dimensionsPerStageIdx = stages.filter((stage) => stage.dimensionId === dimensionId)
-          .length
-        const stageIdx = stages.findIndex((stage) => stage.id === stageId) + 1
-        const routeIdx = Math.ceil(stageIdx / dimensionsPerStageIdx)
-        const discussionURL = makeAppLink(`meet/${meetingId}/estimate/${routeIdx}`)
+        const stageIdx = stages.findIndex((stage) => stage.id === stageId)
+        const discussionURL = makeAppURL(appOrigin, `meet/${meetingId}/estimate/${stageIdx + 1}`)
         const res = await manager.addComment(
           cloudId,
           issueKey,