+++ packages/server/types/modules.d.ts
@@ -6,6 +6,7 @@ declare module '*.svg'
 
 declare const __PRODUCTION__: string
 declare const __STATIC_IMAGES__: string
+declare const __SOCKET_PORT__: string
 interface Window {
   __ACTION__: any
 }